+++ zcash_client_backend/src/wallet.rs
@@ -628,9 +628,12 @@ impl OvkPolicy {
 /// This is implicitly scoped to an account.
 #[derive(Debug, Clone, PartialEq, Eq)]
 #[cfg(feature = "transparent-inputs")]
-pub struct TransparentAddressMetadata {
-    scope: TransparentKeyScope,
-    address_index: NonHardenedChildIndex,
+pub enum TransparentAddressMetadata {
+    Derived {
+        scope: TransparentKeyScope,
+        address_index: NonHardenedChildIndex,
+    },
+    Standalone(secp256k1::PublicKey),
 }
 
 #[cfg(feature = "transparent-inputs")]
@@ -638,17 +641,27 @@ impl TransparentAddressMetadata {
     /// Returns a `TransparentAddressMetadata` in the given scope for the
     /// given address index.
     pub fn new(scope: TransparentKeyScope, address_index: NonHardenedChildIndex) -> Self {
-        Self {
+        Self::Derived {
             scope,
             address_index,
         }
     }
 
-    pub fn scope(&self) -> TransparentKeyScope {
-        self.scope
+    /// Returns the [`TransparentKeyScope`] of the private key from which the address was derived,
+    /// if known. Returns `None` for standalone addresses in the wallet.
+    pub fn scope(&self) -> Option<TransparentKeyScope> {
+        match self {
+            TransparentAddressMetadata::Derived { scope, .. } => Some(*scope),
+            TransparentAddressMetadata::Standalone(_) => None,
+        }
     }
 
-    pub fn address_index(&self) -> NonHardenedChildIndex {
-        self.address_index
+    /// Returns the BIP 44 [`NonHardenedChildIndex`] at which the address was derived, if known.
+    /// Returns `None` for standalone addresses in the wallet.
+    pub fn address_index(&self) -> Option<NonHardenedChildIndex> {
+        match self {
+            TransparentAddressMetadata::Derived { address_index, .. } => Some(*address_index),
+            TransparentAddressMetadata::Standalone(_) => None,
+        }
     }
 }